Updates from Orion2 protocol 2 to bring Anvelina up to version 2.2.3. Updates include:
Changed start address in ASMI_interface to 24'h100000 to match bootloader. This now
allows updating firmware from the application.

Updated to Quartus 20.1. Removed for loops in sdr_send.v, added a resyncronising mechanism
to sdr_send.v such that if the fifo gets out of order (manifests as a high noise floor in panadapter)
it can be reordered due to an extra bit (9 bits) in the fifo data used to show MSB.

Added a dedicated RX slice for the temp_DAC output. This is used for puresignal
and is intended to be more stable and lock in more quickly due to being switched
at 192KHz -vs- 122.88MHz.

Changed SPI clock in Orion_ADC.v from 25% to 50% as some users reported
instability in the 12V reading and the adc78h90 is spec'd @  min %40 / max %60

Added use of Alex1's upper 16 bits for TX relay control. Namely the ANT1-3 bits.
This is to avoid RF appearing on an RX only ANT due to a later arrival of a HP
message telling us to switch the ANT when we are already in TX mode sending RF.
The client program should send this before going to TX.